+++ internal/app/domain/artist.go
@@ -7,42 +7,42 @@ import (
 )
 
 type Artist struct {
-	Id             int    `json:"id" example:"6" db:"id" validate:"min=0,nonnil"`
+	Id             int64  `json:"id" example:"6" db:"id" validate:"min=0,nonnil"`
 	Name           string `json:"name" example:"Imagine Dragons" db:"name" validate:"max=256,nonnil"`
-	CountLikes     int    `json:"countLikes" example:"54" db:"count_likes" validate:"min=0,nonnil"`
-	CountFollowers int    `json:"countFollowers" example:"1001" db:"count_followers" validate:"min=0,nonnil"`
-	CountListening int    `json:"countListening" example:"7654" db:"count_listening" validate:"min=0,nonnil"`
+	CountLikes     int64  `json:"countLikes" example:"54" db:"count_likes" validate:"min=0,nonnil"`
+	CountFollowers int64  `json:"countFollowers" example:"1001" db:"count_followers" validate:"min=0,nonnil"`
+	CountListening int64  `json:"countListening" example:"7654" db:"count_listening" validate:"min=0,nonnil"`
 }
 
 type ArtistRepo interface {
 	Insert(Artist) error
 	Update(Artist) error
-	Delete(int) error
-	SelectByID(int) (*Artist, error)
+	Delete(int64) error
+	SelectByID(int64) (*Artist, error)
 	GetAll() ([]Artist, error)
 	GetPopular() ([]Artist, error)
-	GetLastId() (id int, err error)
+	GetLastId() (id int64, err error)
 	//GetType() reflect.Type
-	GetSize() (int, error)
+	GetSize() (int64, error)
 }
 
 type ArtistDataTransfer struct {
-	Id     int                 `json:"id" example:"1"`
+	Id     int64               `json:"id" example:"1"`
 	Name   string              `json:"name" example:"Mercury"`
 	Cover  string              `json:"cover" example:"assets/artist_1.png"`
-	Likes  int                 `json:"likes" example:"5"`
+	Likes  int64               `json:"likes" example:"5"`
 	Albums []AlbumDataTransfer `json:"albums"`
 }
 
-func (artist *Artist) GetId() int {
+func (artist *Artist) GetId() int64 {
 	return artist.Id
 }
 
-func (artist *Artist) GetIdRef() *int {
+func (artist *Artist) GetIdRef() *int64 {
 	return &artist.Id
 }
 
-func (artist *Artist) SetId(id int) error {
+func (artist *Artist) SetId(id int64) error {
 	artist.Id = id
 	return nil
 }
@@ -51,7 +51,7 @@ func (artist *Artist) Check() error {
 	return validator.Validate(artist)
 }
 
-func (artist *Artist) GetCountListening() int {
+func (artist *Artist) GetCountListening() int64 {
 	return artist.CountListening
 }
