+++ internal/app/tools/utils/randomGenerator.go
@@ -8,9 +8,9 @@ func RandomRune() string {
 	return string('a' + rune(rand.Intn('z'-'a'+1)))
 }
 
-func RandomWord(maxLen int) string {
+func RandomWord(maxLen int64) string {
 	word := ""
-	for i := 0; i < maxLen; i++ {
+	for i := int64(0); i < maxLen; i++ {
 		word += RandomRune()
 	}
 	return word
